In this I have learned about bean definition inheritance is, and how to configure a parent bean definition and have children beans inherit the bean definitions. what happens when abstract value is true
The answers:
bean definition inheritance is same like class inheritance in java. We can define a bean and to my that bean as a parent, we use parent value set to the id of that bean. To make the bean like an interface(abstract) we use abstract value set to true, which makes the bean as an interface.